Add debug mode and kill command
This change adds the debug mode. If debug mode is enabled,
any commands and messages issued to the bot will be ignored
unless they originate from the bot's specified debug text
channel.

To use, specify the following in the bot's .env file:
DEBUG=True
DEBUG_CHANNEL_ID=<The ID of the channel to debug within>

Setting DEBUG=False disables debug mode.

This change also adds the $kill command. This command
causes the bot to log out and thus terminate execution.
This command can only be executed by the owner of the bot.
To specify the owner of the bot, add the following to the
bot's .env file:
OWNER_USER_ID=<The ID of the bot owner>
If the OWNER_USER_ID is omitted from the .env file, the kill
command will have no effect.