Add TriggeredResponse spamming cooldown
This change introduces the models
TriggeredResponseUsageTimestamp
* user: User
	A foreign key to a User model.
* triggered_response: TriggeredResponse
	A foreign key to a TriggeredResponse
* timestamp: DateTime = datetime.now()
	The last time the User used the TriggeredResponse.

Guild
* guild_id: int (primary key)
	Given by discord.
* guild_name: str
	Given by discord.
* triggered_response_cooldown: int = 0

Which represents a discord guild. The triggered_response_cooldown
value is used to determine if a user is allowed to use a
triggered response again. It is applied to all triggered responses
individually. This is a first-attempt at spam filtering.

Also in this change is the ability to modify the triggered
response cooldown of the guild you're currently in, with the command
$pastadelay <integer>. You need manage_message permissions in the
guild in order to use this command.