Add GLOBAL triggered response cooldown for guilds
This change introduces the ability to choose the type of
cooldown used for triggered responses in guilds. There are
two types to choose from:

GLOBAL:
  All triggered responses, text and image, are available to users
  who have not triggered any responses in the last
  guild.triggered_text_cooldown seconds. If a user triggers any
  response, they cannot trigger any additional responses until
  their cooldown expires.
  Example:
  $pastadelay 500
  <activate copy pasta A>
  <activate copy pasta B fails because it hasn't been 500 seconds.>

PER_RESPONSE:
  This type is the original behaviour, where each triggered response
  has a cooldown for a user individually.
  Example:
  $pastadelay 500
  <activate copy pasta A>
  <activate copy pasta B>
  <activate copy pasta A fails because it hasn't been 500 seconds.>

This change also adds a new command, to allow administrators to control
which type of cooldown they want their guild to use: $cooldowntype <type>

Usage:
$cooldowntype <type>
where type is one of "global", "perpasta", not including the quotation marks.
These types correspond to the GLOBAL and PER_RESPONSE cooldown types.

Migrate 0003 to apply model changes.