Add currencies to guilds
This change introduces the concept of guild currencies,
which will be tokens that guild users may exchange for
in-guild content, privileges, etc.

Guild administrators may create, modify, and delete
currencies. Guild administrators and moderators are responsible
for doling tokens out to users manually, until such time
as sambot can perform this function autonomously.

Users will interact with their Wallet via commands. Wallets store
CurrencyAmounts and are one-to-one with users, and thus
when a user is created, a wallet is created for them automatically.

Migrate 0005 to get the model changes. This migration creates
a wallet for any existing users who do not have one.