Add discord notifications when a streamer goes live
Discord members with manage_messages permissions may now
set up twitch notifications for when a streamer goes live.
The notifications that show up in the specified channel
are basically the same as the MEE6 bot.

The command to create such a notification:
$streamnotify <streamer name> #channel-to-notify

Example:
$streamnotify tantooni #general
Will create a notification that is sent to the #general channel
in the guild the command is used, each time the twitch user
whose name matches 'tantooni' goes live.